Add a basic CLI version
