+++ package.json
@@ -1,30 +1,14 @@
 {
   "name": "page-metadata-service",
-  "version": "0.1.0",
   "description": "A RESTful microservice for generating preview summaries of web pages.",
-  "main": "service.js",
-  "scripts": {
-    "test:lint": "eslint .",
-    "test": "test"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/mozilla/page-metadata-service.git"
-  },
-  "keywords": [
-    "page",
-    "summary",
-    "metadata",
-    "rest",
-    "service",
-    "api"
-  ],
+  "version": "0.1.0",
   "author": "Jared Kerim",
-  "license": "MPL-2.0",
+  "bin": {
+    "page-metadata-service": "bin/cli.js"
+  },
   "bugs": {
     "url": "https://github.com/mozilla/page-metadata-service/issues"
   },
-  "homepage": "https://github.com/mozilla/page-metadata-service#readme",
   "dependencies": {
     "body-parser": "^1.15.2",
     "express": "^4.14.0",
@@ -34,6 +18,27 @@
   },
   "devDependencies": {
     "eslint": "^2.13.1",
-    "eslint-plugin-mozilla": "0.0.3"
+    "eslint-plugin-mozilla": "0.0.3",
+    "yargs": "^4.7.1"
+  },
+  "homepage": "https://github.com/mozilla/page-metadata-service#readme",
+  "keywords": [
+    "page",
+    "summary",
+    "metadata",
+    "rest",
+    "service",
+    "api"
+  ],
+  "license": "MPL-2.0",
+  "main": "service.js",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/mozilla/page-metadata-service.git"
+  },
+  "scripts": {
+    "check": "./bin/cli.js",
+    "test": "test",
+    "test:lint": "eslint ."
   }
 }
